requirements.txt
accelerate==0.34.1
pydantic>=2.5,<3.0
torch==2.5.1  # If torch version is changed, also update in docker/rocm-silogen-finetuning-base.Dockerfile
transformers[tokenizers]==4.49.0  # If transformers version is changed, also update in docker/rocm-silogen-finetuning-base.Dockerfile
datasets==3.2.0
trl==0.13.0
mlflow==2.19.0
peft==0.13.0
ruamel.yaml>0.15.0,<1.0
sentencepiece>=0.1.0,<1.0.0
